Add typedef client
